labelversiondetection test merged with existing applicationmonitoring test
