+++ test/sampleapps/sampleapps.go
@@ -8,6 +8,7 @@ import (
 	"github.com/stretchr/testify/require"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
+	"sigs.k8s.io/e2e-framework/klient/k8s/resources"
 	"sigs.k8s.io/e2e-framework/klient/wait"
 	"sigs.k8s.io/e2e-framework/klient/wait/conditions"
 	"sigs.k8s.io/e2e-framework/pkg/envconf"
@@ -48,4 +49,11 @@ func Restart(ctx context.Context, t *testing.T, config *envconf.Config) context.
 			&sampleDeployment, appsv1.DeploymentAvailable, corev1.ConditionTrue)))
 
 	return ctx
+}
+
+func Get(t *testing.T, ctx context.Context, resource *resources.Resources) corev1.PodList {
+	var pods corev1.PodList
+
+	require.NoError(t, resource.WithNamespace(Namespace).List(ctx, &pods))
+	return pods
 }
