+++ test/testdata/application-monitoring/buildlabels-sample-apps.yaml
